separate dictionary for TPOT light
