+++ tpot/tpot.py
@@ -21,8 +21,10 @@ License along with TPOT. If not, see <http://www.gnu.org/licenses/>.
 """
 
 from .base import TPOTBase
-from .config_classifier import classifier_config_dict, classifier_config_dict_lite
-from .config_regressor import regressor_config_dict, regressor_config_dict_lite
+from .config_classifier import classifier_config_dict
+from .config_regressor import regressor_config_dict
+from .config_classifier_lite import classifier_config_dict_lite
+from .config_regressor_lite import regressor_config_dict_lite
 
 
 class TPOTClassifier(TPOTBase):
